+++ ErsatzTV.FFmpeg/Capabilities/FFmpegKnownFilter.cs
@@ -4,6 +4,7 @@ public record FFmpegKnownFilter
 {
     public static readonly FFmpegKnownFilter ScaleNpp = new("scale_npp");
     public static readonly FFmpegKnownFilter TonemapOpenCL = new("tonemap_opencl");
+    public static readonly FFmpegKnownFilter Libplacebo = new("libplacebo");
 
     private FFmpegKnownFilter(string Name) => this.Name = Name;