+++ ErsatzTV.FFmpeg/Capabilities/FFmpegKnownHardwareAcceleration.cs
@@ -8,19 +8,20 @@ public record FFmpegKnownHardwareAcceleration
     public static readonly FFmpegKnownHardwareAcceleration Vaapi = new("vaapi");
     public static readonly FFmpegKnownHardwareAcceleration VideoToolbox = new("videotoolbox");
     public static readonly FFmpegKnownHardwareAcceleration OpenCL = new("opencl");
+    public static readonly FFmpegKnownHardwareAcceleration Vulkan = new("vulkan");
 
     private FFmpegKnownHardwareAcceleration(string Name) => this.Name = Name;
 
     public string Name { get; }
 
     public static IList<string> AllAccels =>
-        new[]
-        {
-            Amf.Name,
-            Cuda.Name,
-            Qsv.Name,
-            Vaapi.Name,
-            VideoToolbox.Name,
-            OpenCL.Name
-        };
+    [
+        Amf.Name,
+        Cuda.Name,
+        Qsv.Name,
+        Vaapi.Name,
+        VideoToolbox.Name,
+        OpenCL.Name,
+        Vulkan.Name
+    ];
 }